+++ src/rules/declaration-colon-newline-after/index.js
@@ -10,6 +10,7 @@ export const ruleName = "declaration-colon-newline-after"
 export const messages = ruleMessages(ruleName, {
   expectedAfter: () => `Expected newline after ":"`,
   rejectedAfter: () => `Unexpected whitespace after ":"`,
+  expectedAfterMultiLine: () => `Unexpected whitespace after ":" with a multi-line value`,
 })
 
 export default function (expectation) {