+++ src/rules/declaration-colon-space-after/index.js
@@ -10,6 +10,7 @@ export const ruleName = "declaration-colon-space-after"
 export const messages = ruleMessages(ruleName, {
   expectedAfter: () => `Expected single space after ":"`,
   rejectedAfter: () => `Unexpected whitespace after ":"`,
+  expectedAfterSingleLine: () => `Expected single space after ":" with a single-line value`,
 })
 
 export default function (expectation) {
@@ -20,6 +21,7 @@ export default function (expectation) {
       possible: [
         "always",
         "never",
+        "always-single-line",
       ],
     })
     if (!validOptions) { return }
@@ -39,20 +41,21 @@ export function declarationColonSpaceChecker({ locationChecker, root, result, ch
 
     for (let i = 0, l = declString.length; i < l; i++) {
       if (declString[i] !== ":") { continue }
-      check(declString, i, decl)
+      locationChecker({
+        source: declString,
+        index: i,
+        lineCheckStr: decl.value,
+        err: m => {
+          report({
+            message: m,
+            node: decl,
+            index: decl.prop.toString().length + 1,
+            result,
+            ruleName: checkedRuleName,
+          })
+        },
+      })
       break
     }
   })
-
-  function check(source, index, node) {
-    locationChecker({ source, index, err: m =>
-      report({
-        message: m,
-        node,
-        index: node.prop.toString().length + 1,
-        result,
-        ruleName: checkedRuleName,
-      }),
-    })
-  }
 }