+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Head
 
+* Fixed: `declaration-colon-newline-after` `"always-multi-line"` warning message.
+* Added: `"always-single-line"` option to `declaration-colon-space-after`.
 * Added: support for declarations directly on root (e.g. Sass variable declarations).
 
 # 1.1.0